complete Authentication

+++ src/main/java/com/shariff/backend/repository/UserRepository.java
@@ -3,10 +3,11 @@ package com.shariff.backend.repository;
 import com.shariff.backend.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-    boolean existsByEmail(String email);
-} 
+    boolean existsByEmail(String email); // To check if email already exists
+}
